Moved logger back to core
