+++ app/component.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/iotaledger/hive.go/app/daemon"
-	"github.com/iotaledger/hive.go/app/logger"
+	"github.com/iotaledger/hive.go/core/logger"
 )
 
 // ComponentParams defines the parameters configuration of a component.
