+++ core/autopeering/selection/common.go
@@ -3,9 +3,9 @@ package selection
 import (
 	"time"
 
-	"github.com/iotaledger/hive.go/app/logger"
 	"github.com/iotaledger/hive.go/core/autopeering/mana"
 	"github.com/iotaledger/hive.go/core/autopeering/peer"
+	"github.com/iotaledger/hive.go/core/logger"
 )
 
 // Default values for the global parameters.
