+++ core/logger/events_test.go
